Reports SSL Handshake errors when they occur
Instead of just reporting a 500 error with some generic text, I thought it might be useful to know when an actual SSL related error occurred. It the idea is to prevent MitM attacks, it might be important to warn users it might be happening.

Currently doesn't drill down the `getCause()` tree very far, just determines if it was an `SSLHandshakeException`.