Fixes the issue of options.followRedirect not overriding globals.followRedirect if options.followRedirect is false.
(options.attr || globals.attr) works well when attr is not boolean. For boolean, if the value of options.attr is set to false (or any falsey value), it will always fall back to globals.attr. In this case, the default value of globals.followRedirect being true, setting options.followRedirect to false does not set followRedirect to false as expected. This commit fixes this issue.